feat: US aid grades
Add a US scale for aid grades.

It treats A# and C# as being part of the same scale, so that different
grade contexts do not need to be assigned to neighbouring routes, e.g.
if there is a grade with grade A2 next to a route with grade C1.

Aid + mandatory free is not covered by this commit.

Resolves: https://github.com/OpenBeta/sandbag/issues/79