test: Add tests of csvtojson
`src/data/csvtojson.ts` is a script used to convert CSV tables of grades
to JSON files used by sandbag at run time.  The script is not tested in
CI, so it is not obvious if it is broken or if someone forgot to run it
while changing grade tables.

Add tests of this script.
Also check that it does not change the corresponding JSON files during
testing.

Fixes: https://github.com/OpenBeta/sandbag/issues/120