fix: Parse YDS slash grades correctly
YDS slash grades were being assigned the score of the first grade. This
was due to `.includes()` not matching the slash.

Change this to correctly identify slash grades.  Refactor a little to
help clarity.
Add test that score range of a slash grade is strict subset of that of
the two grades, and overlaps with both of them.

Add failing tests slash grades composed of non-consecutive grades and
slashes across number boundaries.

Fixes https://github.com/OpenBeta/sandbag/issues/142